requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database - Python 3.13 compatible versions
sqlalchemy>=2.0.25
asyncpg>=0.30.0
psycopg2-binary>=2.9.11  # Python 3.13 compatible
alembic>=1.13.1

# Database Connectors
oracledb>=2.0.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
openpyxl>=3.1.2

# HTTP Client
httpx>=0.26.0
requests>=2.31.0  # ADDED - used by httpHelper

# Validation and Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication - ADDED
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# LLM Integration
litellm>=1.17.9
openai>=1.6.1
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2
